First pass at x-pack support
