+++ lib/puppet/provider/elasticsearch_user/xpack.rb
@@ -1,0 +1,29 @@
+require 'puppet/provider/elastic_parsedfile'
+
+Puppet::Type.type(:elasticsearch_user).provide(
+  :xpack,
+  :parent => Puppet::Provider::ElasticParsedFile
+) do
+  desc "Provider for X-Pack esusers using plain files."
+
+  xpack_config 'users'
+  confine :exists => default_target
+
+  has_feature :manages_passwords
+
+  text_line :comment,
+            :match => %r{^\s*#}
+
+  record_line :parsed,
+              :fields => %w{name hashed_password},
+              :separator => ':',
+              :joiner => ':'
+
+  def self.valid_attr?(klass, attr_name)
+    if klass.respond_to? :parameters
+      klass.parameters.include?(attr_name)
+    else
+      true
+    end
+  end
+end
