+++ lib/puppet/provider/elasticsearch_user_roles/shield.rb
@@ -1,0 +1,11 @@
+require 'puppet/provider/elastic_user_roles'
+
+Puppet::Type.type(:elasticsearch_user_roles).provide(
+  :shield,
+  :parent => Puppet::Provider::ElasticUserRoles
+) do
+  desc "Provider for Shield user roles (parsed file.)"
+
+  shield_config 'users_roles'
+  confine :exists => default_target
+end
